pyproject.toml
[project]
name = "benchmark-pipeline"
version = "0.3.0"
description = "Unified benchmarking framework for ML inference (Torch, ONNX, OpenVINO, TensorRT)."
authors = [
    { name = "Benchmark-pipeline Jasmin F"}
]
requires-python = ">=3.10"
readme = "README.md"
license = { text = "MIT" }

dependencies = [
    # --- Core ---
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "psutil>=5.9",
    "py-cpuinfo",
    "platformdirs",
    "tabulate",
    "numpy>=1.23,<2.3",
    "pandas>=1.5",

    # --- Frameworks (optional by backend) ---
    "torch>=2.0",
    "onnx>=1.15",
    "onnxruntime>=1.17",
    "openvino>=2024.6.0; platform_machine == 'x86_64' and platform_system == 'Linux'",

    "pynvml",

    # --- Plotting ---
    "matplotlib>=3.7",

    # --- Optional metrics tools ---
    "fvcore",
    "ptflops",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "black>=24.0",
    "ruff>=0.3",
    "mypy>=1.0",
    "pre-commit>=3.0"
]

[tool.black]
line-length = 100
target-version = ["py310"]
skip-string-normalization = true

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "B"]
ignore = ["E203", "E266", "E501", "B905"]
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    "venv",
    ".venv",
    "notebooks"
]

[tool.mypy]
python_version = "3.10"
strict = false
ignore_missing_imports = true
warn_unused_ignores = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q"
testpaths = [
    "bench/tests/unit",
    "bench/tests/integration"
]

[tool.coverage.run]
branch = true
source = ["bench"]

[tool.setuptools]
package-dir = { "" = "." }

[tool.setuptools.packages]
find = { include = ["bench*"] }


[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pre-commit]
repos = [
  { repo = "https://github.com/psf/black", rev = "stable", hooks = [{ id = "black" }] },
  { repo = "https://github.com/charliermarsh/ruff-pre-commit", rev = "v0.3.0", hooks = [{ id = "ruff" }] }
]
